Use int as the return value for getopt(3)
char is unsigned by default on arm, which represents -1 as 255

Also add braces to switch statements for clarity.

+++ tests/cmd_pipe_stdout.c
@@ -29,7 +29,7 @@ int main(int argc, char *argv[])
 	output = cmd_pipe_stdout(argv+1, &error_code, &output_size);
 	write(1, output, output_size);
 	fprintf(stderr, "output_size: %d\n", output_size);
-	fprintf(stderr, "strlen: %lu\n", strlen(output));
+	fprintf(stderr, "strlen: %lu\n", (unsigned long)strlen(output));
 	free(output);
 
 	return error_code;
